requirements.txt
numpy
scipy
matplotlib
altair
metaflow
tqdm
filelock
typer
sh
tabulate
fasttext
tomotopy
scikit-learn
gensim
umap-learn
ratelim
python-dotenv
altair_saver
webdriver_manager
db-dtypes
google-cloud-bigquery
google-cloud-bigquery-storage
google-api-core
fsspec==2022.5.0
pandas
selenium==4.2.0
sentence-transformers
networkx
leidenalg
umap-learn